Add Installer.Core shared library and retarget InstallerGui + Tests (#755)
Phase 1 extraction of shared installation logic into Installer.Core class
library. InstallerGui and Installer.Tests now consume the shared library
instead of duplicating code. CLI Installer refactor is next.

New files:
- Installer.Core/InstallationService.cs — all static install/upgrade methods
- Installer.Core/DependencyInstaller.cs — community dependency downloads
- Installer.Core/ScriptProvider.cs — filesystem + embedded resource abstraction
- Installer.Core/Patterns.cs — shared regex patterns ([GeneratedRegex])
- Installer.Core/Models/ — InstallationProgress, ServerInfo, InstallationResult,
  UpgradeInfo, InstallationResultCode (enum mapping CLI exit codes 0-8)

Key changes:
- SQL scripts embedded as assembly resources for future Dashboard integration
- ScriptProvider.FromDirectory() for CLI/GUI, FromEmbeddedResources() for Dashboard
- AutoDiscover() searches filesystem then falls back to embedded
- Comprehensive [DEBUG] logging throughout all methods for GUI diagnostics
- upgrade.txt missing warning (was silently skipped, now logged)
- GenerateSummaryReport gains optional outputDirectory parameter

Retargeted:
- InstallerGui references Installer.Core, old InstallationService.cs deleted
- Installer.Tests targets net8.0 (was net8.0-windows), no WPF dependency
- Tests use ScriptProvider.FromDirectory() instead of raw file paths
